fix mujoco-py

+++ FrankaSim/mvmae_feature_extractor.py
@@ -35,7 +35,7 @@ class MVMAEFeatureExtractor(BaseFeaturesExtractor):
         self.last_mvmae_loss = None
 
     def forward(self, obs_dict: Dict[str, torch.Tensor]) -> torch.Tensor:
-        obs = obs_dict["observation"].to(next(self.mvmae.parameters()).device)
+        obs = obs_dict["observation"].to(next(self.mvmae.parameters()).device).float()
         # obs is (batch, height, width * 2, channels)
         self.mvmae.train()
         out, mask, encoder_nomask_x = self.mvmae(obs)
